moved waqi sensor to air pollutant
